% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfa2.0.R
\name{ar2qspec}
\alias{ar2qspec}
\title{Quantile Spectrum from AR Model of Quantile Series}
\usage{
ar2qspec(fit, freq = NULL)
}
\arguments{
\item{fit}{object of AR model from \code{qser2sar()} or \code{qser2ar()}}

\item{freq}{sequence of frequencies in [0,1) (default = \code{NULL}: all Fourier frequencies)}
}
\value{
a list with the following elements:
  \item{spec}{matrix or array of quantile spectrum/cross-spectrum}
  \item{freq}{sequence of frequencies}
}
\description{
This function computes quantile spectrum/cross-spectrum (QSPEC) from an AR model of quantile series (QSER).
}
