% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{birthweight}
\alias{birthweight}
\title{Birthweight data}
\format{
An object of class \code{data.frame} with 50000 rows and 12 columns.
}
\source{
natality2022us.csv, <https://www.nber.org/research/data/vital-statistics-natality-birth-data>
}
\usage{
data(birthweight)
}
\description{
Infant birth weight data. \code{Precare} and \code{Education} should be treated as factors.
}
\references{
Koenker, R. (2005). Quantile Regression. Cambridge University Press.
}
\keyword{datasets}
