% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{engel}
\alias{engel}
\title{Engel food expenditure data}
\format{
An object of class \code{data.frame} with 235 rows and 2 columns.
}
\usage{
data(engel)
}
\description{
The Engel food expenditure data from the R package \code{quantreg}.
}
\references{
Koenker, R. (2005). Quantile Regression. Cambridge University Press.
}
\keyword{datasets}
