\name{numerical_r}
\alias{numerical_r}
\title{Generates numerical fitnesses from experimental growth curve observations}
\description{
Produces model-free, slope-based fitness phenotypes from a set of timecourse cell density observations (growth curve) generated by Colonyzer, for example.  This function also optionally plots experimental growth curves together with the smoothed version of the data used to generate the numerical phenotypes and a numerical estimate of the slope of the growth curve (rate of change of cell density with time).
}
\usage{
numerical_r(obsdat,mkPlots=FALSE,span=0.3,nBrute=1000,cDiffDelta=0.0001,mlab="")}
\arguments{
  \item{obsdat}{Data frame containing experimental observation of a single growth curve.  This data.frame should contain a column labelled "Expt.Time" and another labelled "Growth".  The output from the colonyzer.read function (filtered to contain data for a single culture) is appropriate.}
  \item{mkPlots}{Boolean (TRUE or FALSE) specifying whether to draw plots showing smoothed version of observed growth curve and numerical estimate of its slope.}
  \item{span}{Strength of smoothing parameter for Loess function}
  \item{nBrute}{Number of evenly spaced synthetic timepoints generated while smoothing data}
  \item{cDiffDelta}{Size of delta (units of time) for generating centre-difference numerical slope estimate.}
  \item{mlab}{Label for plots (only used if mkPlot == TRUE).}
}
\value{
A list containing four values:
\itemize{
\item nr - Numerical estimate of intrinsic growth rate (1/time)
\item nr_t - Time at which maximum slope of log experimental observations occurs
\item mslp - Numerical estimate of maximum slope of growth curve (cell density/time).  Note that this phenotype is not the same as intrinsic growth rate (or exponential growth rate) and so it might not be what you are looking for.
\item mslp_t - Time at which maximum slope of growth curve is observed
}
Optionally (if mkPlot==TRUE) a two-panel figure showing data, smoothed version of data, numerical estimates of slope on both log and linear scale is displayed.  Linearisation of curves about solutions also shown.
}
\keyword{qfa}