\name{qfa-internal} 
\alias{NAtoBlank} 
\alias{bc2st} 
\alias{bctimes} 
\alias{benschopFromSource} 
\alias{buildBenschop} 
\alias{buildGO} 
\alias{colony.fit} 
\alias{colony.info} 
\alias{data.fit} 
\alias{de.fit} 
\alias{dens} 
\alias{drawSel} 
\alias{epiplot}
\alias{getResults}
\alias{getText}
\alias{gethits}
\alias{guess}
\alias{guessNEW}
\alias{index2pos}
\alias{keybd}
\alias{lin.fit}
\alias{lm.epi}
\alias{logdraw}
\alias{makePlot}
\alias{mkPlt}
\alias{mouse}
\alias{na2zero}
\alias{nozero}
\alias{orf2g}
\alias{orf2gdict}
\alias{orffit}
\alias{orfstat}
\alias{orfun}
\alias{posmake}
\alias{ratPlot}
\alias{rcget}
\alias{rowplot}
\alias{slopeFun}
\alias{sterr}
\alias{summarise}
\alias{sumsq}
\alias{targFun}
\alias{tconv}
\alias{testSlope}
\alias{trep}
\alias{typemake}
\alias{varposget}
\alias{globs} 
\alias{getMode}

\title{Internal qfa Functions} 
\description{ 
  Internal qfa functions and environments.  globs is an environment for storing "global" variables.
} 
\details{ 
  These are not to be called by the user. 
} 
\keyword{ internal } 