% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfa4.1.R
\name{qkl.divergence}
\alias{qkl.divergence}
\title{Kullback-Leibler Divergence of Quantile Spectral Estimate}
\usage{
qkl.divergence(y.qper, qspec, sel.f = NULL, sel.tau = NULL)
}
\arguments{
\item{y.qper}{matrix or array of quantile spectral estimate from, e.g., \code{qspec.lw()}}

\item{qspec}{matrix of array of true quantile spectrum (same dimension as \code{y.qper})}

\item{sel.f}{index of selected frequencies for computation (default = \code{NULL}: all frequencies)}

\item{sel.tau}{index of selected quantile levels for computation (default = \code{NULL}: all quantile levels)}
}
\value{
real number of Kullback-Leibler divergence
}
\description{
This function computes Kullback-Leibler divergence (KLD) of quantile spectral estimate.
}
