\name{rod.write}
\alias{rod.write}
\title{Writes a synthetic ROD-like output file to hard-drive.  Deprecated.}
\description{
Takes the data.frame generated by colonyzer.read, discards some columns, rearranges it, and writes data.frame to file in ROD-like format.  This function is
deprecated, and is only provided for backwards compatibility with other deprecated code.
}
\usage{
rod.write(iman,outf)
}
\arguments{
  \item{iman}{Image analysis data.frame as returned by colonyzer.read}
  \item{outf}{Name of output file for writing data to hard-drive.}
}
\value{
Doesn't return anything, but rather writes a tab delimited text file (ROD style format) to hard-drive.
}
\examples{
}
\keyword{qfa}
