\name{showDemo}
\alias{showDemo}
\title{Show Demo}
\description{
Displays source code for demo on screen.
}
\usage{
showDemo(demoname = "telomereCap")
}
\arguments{
  \item{demoname}{Demo to be printed to screen.  Default value "telomerecap" corresponds to the main demo in the QFA package.}
}
\value{
Opens a new window displaying the source code for the demo
}
\keyword{qfa}