% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{yearssn}
\alias{yearssn}
\title{Yearly sunspot numbers}
\format{
An object of class \code{data.frame} with 308 rows and 2 columns.
}
\usage{
data(yearssn)
}
\description{
Sunspot numbers from 1700 to 2007.
}
\references{
Li, T.-H. (2014). Time Series with Mixed Spectra. CRC Press.
}
\keyword{datasets}
