\name{data.fit}
\alias{data.fit}
\title{Fitting generalised logistic model to growth data by least squares (using optim function)}
\description{
Uses the optim function to carry out least squares fit of logistic model to growth data.  
}
\usage{

}
\arguments{

}

\value{

}
\keyword{qfa}