% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfa4.0.R
\name{per}
\alias{per}
\title{Periodogram (PER)}
\usage{
per(y)
}
\arguments{
\item{y}{vector or matrix of time series s (if matrix, nrow(y) = length of time series)}
}
\value{
vector or array of periodogram
}
\description{
This function computes the periodogram or periodogram matrix for univariate or multivariate time series.
}
\examples{
y <- stats::arima.sim(list(order=c(1,0,0), ar=0.5), n=64)
y.per <- per(y)
plot(y.per)
}
