% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfa1.2.R
\name{qsmooth.qper}
\alias{qsmooth.qper}
\title{Quantile Smoothing of Quantile Periodogram or Spectral Estimate}
\usage{
qsmooth.qper(
  qper,
  method = c("gamm", "sp"),
  spar = "GCV",
  n.cores = 1,
  cl = NULL
)
}
\arguments{
\item{qper}{matrix or array of quantile periodogram/cross-periodogram or spectral estimate}

\item{method}{smoothing method: \code{"gamm"} for \code{mgcv::gamm()}, \code{"sp"} for \code{stats::smooth.spline()}}

\item{spar}{smoothing parameter in \code{smooth.spline()} (default = \code{"GCV"})}

\item{n.cores}{number of cores for parallel computing (default = 1)}

\item{cl}{pre-existing cluster for repeated parallel computing (default = \code{NULL})}
}
\value{
matrix or array of quantile-smoothed quantile spectral estimate
}
\description{
This function computes quantile-smoothed version of quantile periodogram/cross-periodogram (QPER) or other quantile spectral estimate.
}
\examples{
y1 <- stats::arima.sim(list(order=c(1,0,0), ar=0.5), n=64)
y2 <- stats::arima.sim(list(order=c(1,0,0), ar=-0.5), n=64)
tau <- seq(0.1,0.9,0.05)
n <- length(y1)
ff <- c(0:(n-1))/n
sel.f <- which(ff > 0 & ff < 0.5)
y.qdft <- qdft(cbind(y1,y2),tau)
qacf <- qdft2qacf(y.qdft)
qper.lw <- qspec.lw(qacf,M=5)$spec
qfa.plot(ff[sel.f],tau,Re(qper.lw[1,1,sel.f,]))
qper.lwqs <- qsmooth.qper(qper.lw,method="sp",spar=0.9)
qfa.plot(ff[sel.f],tau,Re(qper.lwqs[1,1,sel.f,]))
}
